requirements.txt
tensorflow
numpy
streamlit
scikit-learn
pillow